requirements.txt
# Covenant dependencies

# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database + vector store
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.2
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (OPA + outbound)
httpx==0.27.2

# Anthropic SDK
anthropic==0.40.0

# Config
pydantic-settings==2.5.2
python-multipart==0.0.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
